+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/QueryTraverser.scala
@@ -20,6 +20,8 @@ import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectGetADM
 import org.knora.webapi.messages.admin.responder.projectsmessages.ProjectIdentifierADM._
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 
+import org.knora.webapi.messages.util.search.gravsearch.ConstructToConstructTransformer
+
 /**
  * A trait for classes that visit statements and filters in WHERE clauses, accumulating some result.
  *
