+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/SparqlTransformer.scala
@@ -19,89 +19,15 @@ import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 
 import SparqlTransformer._
 
-import SparqlTransformer._
-
 /**
  * Transforms a non-triplestore-specific SELECT for a triplestore that does not have inference enabled (e.g., Fuseki).
  *
  * @param simulateInference `true` if RDFS inference should be simulated using property path syntax.
  */
-class SelectToSelectTransformer(
-  simulateInference: Boolean,
-  sparqlTransformerLive: SparqlTransformerLive,
-  implicit val stringFormatter: StringFormatter
-) extends WhereTransformer {
-
-  /**
-   * Transforms a [[StatementPattern]] in a SELECT's WHERE clause into zero or more statement patterns.
-   *
-   * @param statementPattern the statement to be transformed.
-   * @return the result of the transformation.
-   */
-  def transformStatementInSelect(statementPattern: StatementPattern): Task[Seq[StatementPattern]] =
-    ZIO.succeed(Seq(statementPattern))
-
-  override def transformStatementInWhere(
-    statementPattern: StatementPattern,
-    inputOrderBy: Seq[OrderCriterion],
-    limitInferenceToOntologies: Option[Set[SmartIri]] = None
-  ): Task[Seq[QueryPattern]] =
-    sparqlTransformerLive.transformStatementInWhereForNoInference(
-      statementPattern = statementPattern,
-      simulateInference = simulateInference,
-      limitInferenceToOntologies = limitInferenceToOntologies
-    )
-
-  override def optimiseQueryPatterns(patterns: Seq[QueryPattern]): Task[Seq[QueryPattern]] =
-    ZIO.attempt(moveBindToBeginning(optimiseIsDeletedWithFilter(moveLuceneToBeginning(patterns))))
-
-  override def transformLuceneQueryPattern(luceneQueryPattern: LuceneQueryPattern): Task[Seq[QueryPattern]] =
-    sparqlTransformerLive.transformLuceneQueryPatternForFuseki(luceneQueryPattern)
-
-  /**
-   * Specifies a FROM clause, if needed.
-   *
-   * @return the FROM clause to be used, if any.
-   */
-  def getFromClause: Task[Option[FromClause]] = ZIO.succeed(None)
-
-}
 
 /**
  * Transforms a non-triplestore-specific CONSTRUCT query for a triplestore that does not have inference enabled (e.g., Fuseki).
  */
-class ConstructToConstructTransformer(
-  sparqlTransformerLive: SparqlTransformerLive,
-  implicit val stringFormatter: StringFormatter
-) extends WhereTransformer {
-
-  /**
-   * Transforms a [[StatementPattern]] in a CONSTRUCT clause into zero or more statement patterns.
-   *
-   * @param statementPattern the statement to be transformed.
-   * @return the result of the transformation.
-   */
-  def transformStatementInConstruct(statementPattern: StatementPattern): Task[Seq[StatementPattern]] =
-    ZIO.succeed(Seq(statementPattern))
-
-  override def transformStatementInWhere(
-    statementPattern: StatementPattern,
-    inputOrderBy: Seq[OrderCriterion],
-    limitInferenceToOntologies: Option[Set[SmartIri]] = None
-  ): Task[Seq[QueryPattern]] =
-    sparqlTransformerLive.transformStatementInWhereForNoInference(
-      statementPattern = statementPattern,
-      simulateInference = true,
-      limitInferenceToOntologies = limitInferenceToOntologies
-    )
-
-  override def optimiseQueryPatterns(patterns: Seq[QueryPattern]): Task[Seq[QueryPattern]] =
-    ZIO.attempt(moveBindToBeginning(optimiseIsDeletedWithFilter(moveLuceneToBeginning(patterns))))
-
-  override def transformLuceneQueryPattern(luceneQueryPattern: LuceneQueryPattern): Task[Seq[QueryPattern]] =
-    sparqlTransformerLive.transformLuceneQueryPatternForFuseki(luceneQueryPattern)
-
-}
 
 /**
  * Functions for transforming generated SPARQL.
